fix: refactor SceneLayerEffects to be more generic
