+++ packages/kairos-connection/src/kairos-types/lib-types.ts
@@ -3,3 +3,19 @@ export type ColorRGB = {
 	green: number
 	blue: number
 }
+
+export interface Pos3Df {
+	x: number
+	y: number
+	z: number
+}
+
+export interface Pos2Df {
+	x: number
+	y: number
+}
+
+export interface Pos2D {
+	x: number
+	y: number
+}
