+++ packages/kairos-connection/src/kairos-types/main.ts
@@ -1,5 +1,6 @@
 export * from './lib-types.js'
 export * from './clip-player.js'
+export * from './effects.js'
 export * from './macro.js'
 export * from './media.js'
 export * from './scene.js'
