+++ packages/kairos-connection/src/object-encoding/index.ts
@@ -1,10 +1,10 @@
 export * from './types.js'
 export * from './scene.js'
 export * from './scene-layer.js'
-export * from './scene-layer-effects-basic.js'
-export * from './scene-layer-effects-key.js'
-export * from './scene-layer-effects-color.js'
-export * from './scene-layer-effects-visual.js'
+export * from './effects-basic.js'
+export * from './effects-key.js'
+export * from './effects-color.js'
+export * from './effects-visual.js'
 export * from './scene-transitions.js'
 export * from './scene-snapshots.js'
 export * from './macros.js'
